controller 에러핸들링 적용

+++ src/controllers/userController.js
@@ -17,20 +17,31 @@ const signUp = catchAsync(async (req, res) => {
 const signIn = catchAsync(async (req, res) => {
   const { account, password } = req.body;
 
+  if (!account || !password) {
+    return res.status(400).json({ message: "KEY_ERROR" });
+  }
+
   const accessToken = await userService.signIn(account, password);
   return res.status(200).json({ accessToken: accessToken });
 });
 
 const getFavorites = catchAsync(async (req, res) => {
   const account = req.user;
-  const favorites = await userService.getFavorites(account);
+  if (!account) {
+    return res.status(400).json({ message: "KEY_ERROR" });
+  }
 
+  const favorites = await userService.getFavorites(account);
   return res.status(200).json(favorites);
 });
 
 const addFavorites = catchAsync(async (req, res) => {
   const account = req.user;
   const cafe_id = req.params.cafeId;
+
+  if (!account || !cafe_id) {
+    return res.status(400).json({ message: "KEY_ERROR" });
+  }
 
   await userService.addFavorites(account, cafe_id);
   return res.status(201).json({
@@ -41,6 +52,10 @@ const addFavorites = catchAsync(async (req, res) => {
 const deleteFavorites = catchAsync(async (req, res) => {
   const account = req.user;
   const cafe_id = req.params.cafeId;
+
+  if (!account || !cafe_id) {
+    return res.status(400).json({ message: "KEY_ERROR" });
+  }
 
   await userService.deleteFavorites(account, cafe_id);
   return res.status(204).send();
